refactor: Update 'no_qualified' enum value in QualificationOption.php, QualificationFactory.php, and create_qualifications_table.php
Update the 'NotQualified' enum value to 'no_qualified' in QualificationOption.php, QualificationFactory.php, and create_qualifications_table.php. This change improves consistency and clarity by using a more descriptive term to represent the qualification status of a user. The 'no_qualified' value is now used to indicate that a user is not qualified for a certain qualification option. This modification enhances the codebase by providing a more intuitive and meaningful representation of qualification statuses.