refactor: Update ClientResource to use clientType name instead of priority
This commit modifies the ClientResource class to use the name property of the clientType instead of the priority property. By doing so, the client_type field in the toArray method now returns the name of the client type instead of its priority. This change improves the clarity and accuracy of the data being returned by the resource.